positionCode to read by engine

+++ src/main/java/chessEngine/chess/piece/infiniteRangePiece/bishop/BlackBishop.java
@@ -6,7 +6,7 @@ public class BlackBishop extends Bishop {
 
     private final PieceColor pieceColor = PieceColor.BLACK;
 
-    public BlackBishop(int height, int width) {
+    public BlackBishop(byte height, byte width) {
         super(height, width);
     }
 }
